requirements.txt
requests
python-dotenv
streamlit
pandas
openai